docker-elasticsearch/requirements.txt
Click==7.1.2
elasticsearch==7.14.2
elasticsearch-curator==5.8.4